Created InsertClientActivity and did some refactoring. When pressing on a group in NewClientActivity, will take to InserClientActivity.
