+++ app/src/main/java/rainvagel/healthreporter/ClientClasses/InsertClientActivity.java
@@ -1,0 +1,15 @@
+package rainvagel.healthreporter.ClientClasses;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import rainvagel.healthreporter.R;
+
+public class InsertClientActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_insert_client);
+    }
+}
