+++ app/src/main/res/layout/activity_client.xml
@@ -6,7 +6,7 @@
 
     android:layout_height="match_parent"
 
-    tools:context="rainvagel.healthreporter.ClientActivity">
+    tools:context="rainvagel.healthreporter.ClientClasses.ClientActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
