+++ app/src/main/res/layout/activity_new_client.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="rainvagel.healthreporter.NewClientActivity">
+    tools:context="rainvagel.healthreporter.ClientClasses.NewClientActivity">
     <!--<FrameLayout-->
         <!--android:layout_width="match_parent"-->
         <!--android:layout_height="wrap_content">-->
