+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context=".ClientActivity">
+    xmlns:tools="http://schemas.android.com/tools" tools:context=".ClientClasses.ClientActivity">
     <item android:id="@+id/action_settings" android:title="@string/app_name"
         android:orderInCategory="100" app:showAsAction="never" />
